+++ ChangeLog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add struct `ODSExtrinsicCalibrationCorrectionV1` in Deserialize module
+- Support for CMake 4.0
 
 ## 1.6.7 - 2025-03-28
 - Update the CLI documentation